requirements.txt
alembic==1.10.1
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
SQLAlchemy==2.0.3
typing_extensions==4.5.0
Werkzeug==2.2.3
